перенес проверку загрузки данных с сервера на уровень выше в app устранил проблему с тем, что не открывались страницы напрямую через адресную строку
