+++ src/components/app/app.jsx
@@ -1,49 +1,32 @@
 import React from 'react';
-import {BrowserRouter, Switch, Route} from 'react-router-dom';
-import Main from 'components/main/main';
-import SignIn from 'components/sign-in/sign-in';
-import MyList from 'components/my-list/my-list';
-import Movie from 'components/movie/movie';
-import Review from 'components/review/review';
-import Player from 'components/player/player';
+import PropTypes from 'prop-types';
+import {connect} from 'react-redux';
+import Routing from 'components/routing/routing';
+import {init} from 'store/api-action';
 
 
-const App = () => {
+class App extends React.PureComponent {
+  componentDidMount() {
+    this.props.initAction();
+  }
 
-  return (
-    <BrowserRouter>
-      <Switch>
-        <Route exact path='/'
-          render={({history}) => (
-            <Main onMovieCardClick={(id) => history.push(`/films/${id}`)} />
-          )}
-        />
-        <Route exact path='/login'>
-          <SignIn />
-        </Route>
-        <Route exact path='/mylist'
-          render={({history}) => (
-            <MyList onMovieCardClick={(id) => history.push(`/films/${id}`)} />
-          )}
-        />
-        <Route exact path='/films/:id'
-          render={({history, match}) => (
-            <Movie match={match}
-              onMovieCardClick={(id) => history.push(`/films/${id}`)} />
-          )}
-        />
-        <Route exact path='/films/:id/review'
-          render={({match}) => (
-            <Review match={match}/>
-          )}
-        />
-        <Route exact path='/player/:id'>
-          <Player />
-        </Route>
-      </Switch>
-    </BrowserRouter>
-  );
+  render() {
+    return this.props.isApplicationReady && <Routing />;
+  }
+}
+
+App.propTypes = {
+  isApplicationReady: PropTypes.bool.isRequired,
+  initAction: PropTypes.func.isRequired,
+};
+
+const mapStateToProps = ({moviesState}) => ({
+  isApplicationReady: moviesState.isApplicationReady,
+});
+
+const mapDispatchToProps = {
+  initAction: init,
 };
 
 
-export default App;
+export default connect(mapStateToProps, mapDispatchToProps)(App);
